requirements.txt
flask
numpy
opencv-python
tensorflow-gpu